+++ src/Sulu/Bundle/RouteBundle/Entity/RouteRepository.php
@@ -38,4 +38,24 @@ class RouteRepository extends EntityRepository implements RouteRepositoryInterfa
             return;
         }
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function findByEntity($entityClass, $entityId, $locale)
+    {
+        $query = $this->createQueryBuilder('entity')
+            ->andWhere('entity.entityClass = :entityClass')
+            ->andWhere('entity.entityId = :entityId')
+            ->andWhere('entity.locale = :locale')
+            ->andWhere('entity.history = false')
+            ->getQuery()
+            ->setParameters(['entityClass' => $entityClass, 'entityId' => $entityId, 'locale' => $locale]);
+
+        try {
+            return $query->getSingleResult();
+        } catch (NoResultException $e) {
+            return;
+        }
+    }
 }