+++ src/Sulu/Bundle/RouteBundle/Generator/CannotEvaluateTokenException.php
@@ -34,7 +34,7 @@ class CannotEvaluateTokenException extends \Exception
     public function __construct($token, $entity, \Exception $previous)
     {
         parent::__construct(
-            sprintf('Cannot evaluate toeken "%s" for entity with type "%s"', $token, get_class($entity)),
+            sprintf('Cannot evaluate token "%s" for entity with type "%s"', $token, get_class($entity)),
             0,
             $previous
         );