+++ src/Sulu/Bundle/RouteBundle/Manager/RouteManager.php
@@ -11,9 +11,7 @@
 
 namespace Sulu\Bundle\RouteBundle\Manager;
 
-use Sulu\Bundle\RouteBundle\Entity\RouteRepositoryInterface;
-use Sulu\Bundle\RouteBundle\Exception\MissingClassMappingConfigurationException;
-use Sulu\Bundle\RouteBundle\Generator\RouteGeneratorInterface;
+use Sulu\Bundle\RouteBundle\Generator\ChainRouteGeneratorInterface;
 use Sulu\Bundle\RouteBundle\Model\RoutableInterface;
 
 /**
@@ -22,19 +20,9 @@ use Sulu\Bundle\RouteBundle\Model\RoutableInterface;
 class RouteManager implements RouteManagerInterface
 {
     /**
-     * @var array
+     * @var ChainRouteGeneratorInterface
      */
-    private $mappings;
-
-    /**
-     * @var RouteGeneratorInterface
-     */
-    private $routeGenerators;
-
-    /**
-     * @var RouteRepositoryInterface
-     */
-    private $routeRepository;
+    private $chainRouteGenerator;
 
     /**
      * @var ConflictResolverInterface
@@ -42,21 +30,15 @@ class RouteManager implements RouteManagerInterface
     private $conflictResolver;
 
     /**
-     * @param RouteGeneratorInterface[] $routeGenerators
-     * @param RouteRepositoryInterface $routeRepository
+     * @param ChainRouteGeneratorInterface $chainRouteGenerator
      * @param ConflictResolverInterface $conflictResolver
-     * @param array $mappings
      */
     public function __construct(
-        array $routeGenerators,
-        RouteRepositoryInterface $routeRepository,
-        ConflictResolverInterface $conflictResolver,
-        array $mappings
+        ChainRouteGeneratorInterface $chainRouteGenerator,
+        ConflictResolverInterface $conflictResolver
     ) {
-        $this->routeGenerators = $routeGenerators;
-        $this->routeRepository = $routeRepository;
+        $this->chainRouteGenerator = $chainRouteGenerator;
         $this->conflictResolver = $conflictResolver;
-        $this->mappings = $mappings;
     }
 
     /**
@@ -68,47 +50,11 @@ class RouteManager implements RouteManagerInterface
             throw new RouteAlreadyCreatedException($entity);
         }
 
-        $config = $this->getClassMappingConfiguration(get_class($entity));
-
-        if (null === $path) {
-            $path = $this->routeGenerators[$config['generator']]->generate($entity, $config['options']);
-        }
-
-        $route = $this->routeRepository->createNew()
-            ->setPath($path)
-            ->setEntityClass(get_class($entity))
-            ->setEntityId($entity->getId())
-            ->setLocale($entity->getLocale());
-
+        $route = $this->chainRouteGenerator->generate($entity, $path);
         $route = $this->conflictResolver->resolve($route);
         $entity->setRoute($route);
 
         return $route;
-    }
-
-    /**
-     * Get class mapping configuration by class name or inheritance chain.
-     *
-     * @param string $className
-     *
-     * @return array
-     *
-     * @throws MissingClassMappingConfigurationException
-     */
-    protected function getClassMappingConfiguration($className)
-    {
-        if (array_key_exists($className, $this->mappings)) {
-            return $this->mappings[$className];
-        }
-
-        $reflection = new \ReflectionClass($className);
-        while ($reflection = $reflection->getParentClass()) {
-            if (array_key_exists($reflection->getName(), $this->mappings)) {
-                return $this->mappings[$reflection->getName()];
-            }
-        }
-
-        throw new MissingClassMappingConfigurationException($className, array_keys($this->mappings));
     }
 
     /**
@@ -120,21 +66,11 @@ class RouteManager implements RouteManagerInterface
             throw new RouteNotCreatedException($entity);
         }
 
-        $config = $this->getClassMappingConfiguration(get_class($entity));
-
-        if (null === $path) {
-            $path = $this->routeGenerators[$config['generator']]->generate($entity, $config['options']);
-        }
-
-        if ($path === $entity->getRoute()->getPath()) {
+        $route = $this->chainRouteGenerator->generate($entity, $path);
+        if ($route->getPath() === $entity->getRoute()->getPath()) {
             return $entity->getRoute();
         }
 
-        $route = $this->routeRepository->createNew()
-            ->setPath($path)
-            ->setEntityClass(get_class($entity))
-            ->setEntityId($entity->getId())
-            ->setLocale($entity->getLocale());
         $route = $this->conflictResolver->resolve($route);
 
         // path haven't changed after conflict resolving