+++ src/Sulu/Bundle/RouteBundle/Tests/Unit/DependencyInjection/RouteGeneratorCompilerPassTest.php
@@ -54,7 +54,7 @@ class RouteGeneratorCompilerPassTest extends \PHPUnit_Framework_TestCase
 
         $definition = $this->prophesize(Definition::class);
         $definition->replaceArgument(
-            0,
+            1,
             Argument::that(
                 function ($argument) use ($generatorAlias, $serviceId) {
                     return 1 === count($argument) && $argument[$generatorAlias]->__toString() === $serviceId;
@@ -112,7 +112,7 @@ class RouteGeneratorCompilerPassTest extends \PHPUnit_Framework_TestCase
 
         $definition = $this->prophesize(Definition::class);
         $definition->replaceArgument(
-            0,
+            1,
             Argument::that(
                 function ($argument) use ($serviceId) {
                     return 0 === count($argument);