+++ src/Sulu/Bundle/RouteBundle/Tests/Unit/Manager/RouteManagerTest.php
@@ -11,9 +11,9 @@
 
 namespace Sulu\Bundle\RouteBundle\Tests\Unit\Manager;
 
+use Sulu\Bundle\RouteBundle\Entity\Route;
 use Sulu\Bundle\RouteBundle\Entity\RouteRepositoryInterface;
-use Sulu\Bundle\RouteBundle\Exception\MissingClassMappingConfigurationException;
-use Sulu\Bundle\RouteBundle\Generator\RouteGeneratorInterface;
+use Sulu\Bundle\RouteBundle\Generator\ChainRouteGeneratorInterface;
 use Sulu\Bundle\RouteBundle\Manager\ConflictResolverInterface;
 use Sulu\Bundle\RouteBundle\Manager\RouteAlreadyCreatedException;
 use Sulu\Bundle\RouteBundle\Manager\RouteManager;
@@ -24,14 +24,9 @@ use Sulu\Bundle\RouteBundle\Model\RouteInterface;
 class RouteManagerTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @var array
+     * @var ChainRouteGeneratorInterface
      */
-    private $mappings;
-
-    /**
-     * @var RouteGeneratorInterface
-     */
-    private $routeGenerator;
+    private $chainRouteGenerator;
 
     /**
      * @var RouteRepositoryInterface
@@ -56,59 +51,26 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
     protected function setUp()
     {
         $this->entity = $this->prophesize(RoutableInterface::class);
-        $this->mappings = [
-            get_class($this->entity->reveal()) => [
-                'generator' => 'route_generator',
-                'options' => [
-                    'route_schema' => '/{title}',
-                ],
-            ],
-            TestRoutable::class => [
-                'generator' => 'route_generator',
-                'options' => [
-                    'route_schema' => '/{title}',
-                ],
-            ],
-        ];
-
-        $this->routeGenerator = $this->prophesize(RouteGeneratorInterface::class);
-        $this->routeRepository = $this->prophesize(RouteRepositoryInterface::class);
+        $this->chainRouteGenerator = $this->prophesize(ChainRouteGeneratorInterface::class);
         $this->conflictResolver = $this->prophesize(ConflictResolverInterface::class);
 
         $this->manager = new RouteManager(
-            ['route_generator' => $this->routeGenerator->reveal()],
-            $this->routeRepository->reveal(),
-            $this->conflictResolver->reveal(),
-            $this->mappings
+            $this->chainRouteGenerator->reveal(),
+            $this->conflictResolver->reveal()
         );
     }
 
     public function testCreate()
     {
         $route = $this->prophesize(RouteInterface::class);
-        $route->setPath('/test')->shouldBeCalled()->willReturn($route->reveal());
-        $route->setEntityClass(get_class($this->entity->reveal()))->shouldBeCalled()->willReturn($route->reveal());
-        $route->setEntityId('1')->shouldBeCalled()->willReturn($route->reveal());
-        $route->setLocale('de')->shouldBeCalled()->willReturn($route->reveal());
 
-        $this->entity->getId()->willReturn('1');
-        $this->entity->getLocale()->willReturn('de');
         $this->entity->getRoute()->willReturn(null);
         $this->entity->setRoute($route->reveal())->shouldBeCalled();
 
-        $this->routeGenerator->generate($this->entity->reveal(), ['route_schema' => '/{title}'])->willReturn('/test');
-        $this->routeRepository->createNew()->willReturn($route->reveal());
+        $this->chainRouteGenerator->generate($this->entity->reveal(), null)->willReturn($route->reveal());
         $this->conflictResolver->resolve($route->reveal())->willReturn($route->reveal());
 
         $this->assertEquals($route->reveal(), $this->manager->create($this->entity->reveal()));
-    }
-
-    public function testCreateNoMapping()
-    {
-        $this->setExpectedException(MissingClassMappingConfigurationException::class);
-        $entity = $this->prophesize(RoutableInterface::class);
-
-        $this->manager->create($entity->reveal());
     }
 
     public function testCreateInheritMapping()
@@ -116,13 +78,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
         $entity = new TestRoutableProxy();
 
         $route = $this->prophesize(RouteInterface::class);
-        $route->setPath('/test')->shouldBeCalled()->willReturn($route->reveal());
-        $route->setEntityClass(get_class($entity))->shouldBeCalled()->willReturn($route->reveal());
-        $route->setEntityId('1')->shouldBeCalled()->willReturn($route->reveal());
-        $route->setLocale('de')->shouldBeCalled()->willReturn($route->reveal());
-
-        $this->routeGenerator->generate($entity, ['route_schema' => '/{title}'])->willReturn('/test');
-        $this->routeRepository->createNew()->willReturn($route->reveal());
+        $this->chainRouteGenerator->generate($entity, null)->willReturn($route->reveal());
         $this->conflictResolver->resolve($route->reveal())->willReturn($route->reveal());
 
         $this->assertEquals($route->reveal(), $this->manager->create($entity));
@@ -132,18 +88,11 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
     public function testCreateWithRoutePath()
     {
         $route = $this->prophesize(RouteInterface::class);
-        $route->setPath('/test')->shouldBeCalled()->willReturn($route->reveal());
-        $route->setEntityClass(get_class($this->entity->reveal()))->shouldBeCalled()->willReturn($route->reveal());
-        $route->setEntityId('1')->shouldBeCalled()->willReturn($route->reveal());
-        $route->setLocale('de')->shouldBeCalled()->willReturn($route->reveal());
 
-        $this->entity->getId()->willReturn('1');
-        $this->entity->getLocale()->willReturn('de');
         $this->entity->getRoute()->willReturn(null);
         $this->entity->setRoute($route->reveal())->shouldBeCalled();
 
-        $this->routeGenerator->generate($this->entity->reveal(), ['route_schema' => '/{title}'])->shouldNotBeCalled();
-        $this->routeRepository->createNew()->willReturn($route->reveal());
+        $this->chainRouteGenerator->generate($this->entity->reveal(), '/test')->willReturn($route->reveal());
         $this->conflictResolver->resolve($route->reveal())->willReturn($route->reveal());
 
         $this->assertEquals($route->reveal(), $this->manager->create($this->entity->reveal(), '/test'));
@@ -163,20 +112,9 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
     public function testCreateWithConflict()
     {
         $route = $this->prophesize(RouteInterface::class);
-        $route->setPath('/test')->shouldBeCalled()->willReturn($route->reveal());
-        $route->setEntityClass(get_class($this->entity->reveal()))->shouldBeCalled()->willReturn($route->reveal());
-        $route->setEntityId('1')->shouldBeCalled()->willReturn($route->reveal());
-        $route->setLocale('de')->shouldBeCalled()->willReturn($route->reveal());
-
         $conflict = $this->prophesize(RouteInterface::class);
 
-        $this->entity->getId()->willReturn('1');
-        $this->entity->getLocale()->willReturn('de');
-        $this->entity->getRoute()->willReturn(null);
-        $this->entity->setRoute($conflict->reveal())->shouldBeCalled();
-
-        $this->routeGenerator->generate($this->entity->reveal(), ['route_schema' => '/{title}'])->willReturn('/test');
-        $this->routeRepository->createNew()->willReturn($route->reveal());
+        $this->chainRouteGenerator->generate($this->entity->reveal(), null)->willReturn($route->reveal());
         $this->conflictResolver->resolve($route->reveal())->willReturn($conflict->reveal());
 
         $this->assertEquals($conflict->reveal(), $this->manager->create($this->entity->reveal()));
@@ -193,17 +131,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
         $this->entity->getRoute()->willReturn($route->reveal());
 
         $newRoute = $this->prophesize(RouteInterface::class);
-        $newRoute->setEntityClass(get_class($this->entity->reveal()))->shouldBeCalled()
-            ->willReturn($newRoute->reveal());
-        $newRoute->setPath('/test-2')->shouldBeCalled()->will(
-            function () use ($newRoute) {
-                $newRoute->getPath()->willReturn('/test-2');
-
-                return $newRoute->reveal();
-            }
-        );
-        $newRoute->setEntityId('1')->shouldBeCalled()->willReturn($newRoute->reveal());
-        $newRoute->setLocale('de')->shouldBeCalled()->willReturn($newRoute->reveal());
+        $newRoute->getPath()->willReturn('/test-2');
         $newRoute->addHistory($route->reveal())->shouldBeCalled()->willReturn($newRoute->reveal());
 
         $route->setTarget($newRoute->reveal())->shouldBeCalled()->willReturn($route->reveal());
@@ -211,8 +139,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
 
         $this->entity->setRoute($newRoute->reveal())->shouldBeCalled();
 
-        $this->routeGenerator->generate($this->entity->reveal(), ['route_schema' => '/{title}'])->shouldNotBeCalled();
-        $this->routeRepository->createNew()->willReturn($newRoute->reveal());
+        $this->chainRouteGenerator->generate($this->entity->reveal(), '/test-2')->willReturn($newRoute->reveal());
         $this->conflictResolver->resolve($newRoute->reveal())->shouldBeCalled()->willReturn($newRoute->reveal());
 
         $this->assertEquals($newRoute->reveal(), $this->manager->update($this->entity->reveal(), '/test-2'));
@@ -227,27 +154,16 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
         $entity = new TestRoutableProxy($route->reveal());
 
         $newRoute = $this->prophesize(RouteInterface::class);
-        $newRoute->setEntityClass(get_class($entity))->shouldBeCalled()
-            ->willReturn($newRoute->reveal());
-        $newRoute->setPath('/test-2')->shouldBeCalled()->will(
-            function () use ($newRoute) {
-                $newRoute->getPath()->willReturn('/test-2');
-
-                return $newRoute->reveal();
-            }
-        );
-        $newRoute->setEntityId('1')->shouldBeCalled()->willReturn($newRoute->reveal());
-        $newRoute->setLocale('de')->shouldBeCalled()->willReturn($newRoute->reveal());
+        $newRoute->getPath()->willReturn('/test-2');
         $newRoute->addHistory($route->reveal())->shouldBeCalled()->willReturn($newRoute->reveal());
 
         $route->setTarget($newRoute->reveal())->shouldBeCalled()->willReturn($route->reveal());
         $route->getHistories()->willReturn([]);
 
-        $this->routeGenerator->generate($entity, ['route_schema' => '/{title}'])->shouldNotBeCalled();
-        $this->routeRepository->createNew()->willReturn($newRoute->reveal());
+        $this->chainRouteGenerator->generate($entity, null)->willReturn($newRoute->reveal());
         $this->conflictResolver->resolve($newRoute->reveal())->shouldBeCalled()->willReturn($newRoute->reveal());
 
-        $this->assertEquals($newRoute->reveal(), $this->manager->update($entity, '/test-2'));
+        $this->assertEquals($newRoute->reveal(), $this->manager->update($entity));
         $this->assertEquals($newRoute->reveal(), $entity->getRoute());
     }
 
@@ -262,17 +178,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
         $this->entity->getRoute()->willReturn($route->reveal());
 
         $newRoute = $this->prophesize(RouteInterface::class);
-        $newRoute->setEntityClass(get_class($this->entity->reveal()))->shouldBeCalled()
-            ->willReturn($newRoute->reveal());
-        $newRoute->setPath('/test-2')->shouldBeCalled()->will(
-            function () use ($newRoute) {
-                $newRoute->getPath()->willReturn('/test-2');
-
-                return $newRoute->reveal();
-            }
-        );
-        $newRoute->setEntityId('1')->shouldBeCalled()->willReturn($newRoute->reveal());
-        $newRoute->setLocale('de')->shouldBeCalled()->willReturn($newRoute->reveal());
+        $newRoute->getPath()->willReturn('/test-2');
 
         $conflict = $this->prophesize(RouteInterface::class);
         $conflict->addHistory($route->reveal())->shouldBeCalled()->willReturn($conflict->reveal());
@@ -283,8 +189,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
 
         $this->entity->setRoute($conflict->reveal())->shouldBeCalled();
 
-        $this->routeGenerator->generate($this->entity->reveal(), ['route_schema' => '/{title}'])->willReturn('/test-2');
-        $this->routeRepository->createNew()->willReturn($newRoute->reveal());
+        $this->chainRouteGenerator->generate($this->entity->reveal(), null)->willReturn($newRoute->reveal());
         $this->conflictResolver->resolve($newRoute->reveal())->shouldBeCalled()->willReturn($conflict->reveal());
 
         $this->assertEquals($conflict->reveal(), $this->manager->update($this->entity->reveal()));
@@ -301,17 +206,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
         $this->entity->getRoute()->willReturn($route->reveal());
 
         $newRoute = $this->prophesize(RouteInterface::class);
-        $newRoute->setEntityClass(get_class($this->entity->reveal()))->shouldBeCalled()
-            ->willReturn($newRoute->reveal());
-        $newRoute->setPath('/test-2')->shouldBeCalled()->will(
-            function () use ($newRoute) {
-                $newRoute->getPath()->willReturn('/test-2');
-
-                return $newRoute->reveal();
-            }
-        );
-        $newRoute->setEntityId('1')->shouldBeCalled()->willReturn($newRoute->reveal());
-        $newRoute->setLocale('de')->shouldBeCalled()->willReturn($newRoute->reveal());
+        $newRoute->getPath()->willReturn('/test-2');
         $newRoute->addHistory($route->reveal())->shouldBeCalled()->willReturn($newRoute->reveal());
 
         $historyRoute1 = $this->prophesize(RouteInterface::class);
@@ -329,8 +224,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
 
         $this->entity->setRoute($newRoute->reveal())->shouldBeCalled();
 
-        $this->routeGenerator->generate($this->entity->reveal(), ['route_schema' => '/{title}'])->willReturn('/test-2');
-        $this->routeRepository->createNew()->willReturn($newRoute->reveal());
+        $this->chainRouteGenerator->generate($this->entity->reveal(), null)->willReturn($newRoute->reveal());
         $this->conflictResolver->resolve($newRoute->reveal())->shouldBeCalled()->willReturn($newRoute->reveal());
 
         $this->assertEquals($newRoute->reveal(), $this->manager->update($this->entity->reveal()));
@@ -347,17 +241,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
         $this->entity->getRoute()->willReturn($route->reveal());
 
         $newRoute = $this->prophesize(RouteInterface::class);
-        $newRoute->setEntityClass(get_class($this->entity->reveal()))->shouldBeCalled()
-            ->willReturn($newRoute->reveal());
-        $newRoute->setPath('/test-2')->shouldBeCalled()->will(
-            function () use ($newRoute) {
-                $newRoute->getPath()->willReturn('/test-2');
-
-                return $newRoute->reveal();
-            }
-        );
-        $newRoute->setEntityId('1')->shouldBeCalled()->willReturn($newRoute->reveal());
-        $newRoute->setLocale('de')->shouldBeCalled()->willReturn($newRoute->reveal());
+        $newRoute->getPath()->willReturn('/test-2');
 
         $historyRoute1 = $this->prophesize(RouteInterface::class);
         $historyRoute1->removeTarget()->shouldBeCalled()->willReturn($historyRoute1->reveal());
@@ -370,9 +254,7 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
 
         $this->entity->setRoute($historyRoute1->reveal())->shouldBeCalled();
 
-        $this->routeGenerator->generate($this->entity->reveal(), ['route_schema' => '/{title}'])
-            ->willReturn('/test-2');
-        $this->routeRepository->createNew()->willReturn($newRoute->reveal());
+        $this->chainRouteGenerator->generate($this->entity->reveal(), null)->willReturn($newRoute->reveal());
         $this->conflictResolver->resolve($newRoute->reveal())->shouldBeCalled()->willReturn($historyRoute1->reveal());
 
         $this->manager->update($this->entity->reveal());
@@ -385,7 +267,8 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
 
         $this->entity->getRoute()->willReturn($route->reveal());
 
-        $this->routeGenerator->generate($this->entity->reveal(), ['route_schema' => '/{title}'])->willReturn('/test');
+        $this->chainRouteGenerator->generate($this->entity->reveal(), null)
+            ->willReturn(new Route('/test'));
 
         $this->assertEquals($route->reveal(), $this->manager->update($this->entity->reveal()));
     }
@@ -397,18 +280,9 @@ class RouteManagerTest extends \PHPUnit_Framework_TestCase
         $this->entity->getRoute()->willReturn(null);
         $this->entity->getId()->willReturn('1');
 
-        $this->routeGenerator->generate($this->entity->reveal(), '/{title}')->willReturn('/test');
+        $this->chainRouteGenerator->generate($this->entity->reveal())->willReturn('/test');
 
         $this->manager->update($this->entity->reveal());
-    }
-
-    public function testUpdateNoMapping()
-    {
-        $this->setExpectedException(MissingClassMappingConfigurationException::class);
-        $entity = $this->prophesize(RoutableInterface::class);
-        $entity->getRoute()->willReturn($this->prophesize(RouteInterface::class)->reveal());
-
-        $this->manager->update($entity->reveal());
     }
 }